cart function fully works

+++ lib/models/cart.dart
@@ -4,8 +4,8 @@ class Cart {
   late final int? id;
   final String? productId;
   final String? productName;
-  final int? initialPrice;
-  final int? productPrice;
+  final double? initialPrice;
+  final double? productPrice;
   final ValueNotifier<int>? quantity;
   final String? unitTag;
   final String? image;
